+++ report/util.go
@@ -19,11 +19,19 @@ package report
 
 import (
 	"bytes"
+	"encoding/json"
 	"fmt"
+	"io/ioutil"
+	"os"
+	"path/filepath"
+	"regexp"
+	"sort"
 	"strings"
+	"time"
 
 	"github.com/future-architect/vuls/config"
 	"github.com/future-architect/vuls/models"
+	"github.com/future-architect/vuls/util"
 	"github.com/gosuri/uitable"
 )
 
@@ -515,4 +523,257 @@ func formatOneChangelog(p models.Package) string {
 	}
 	buf = append(buf, packVer, delim.String(), clog)
 	return strings.Join(buf, "\n")
+}
+
+func needToRefreshCve(r models.ScanResult) bool {
+	if r.Lang != config.Conf.Lang {
+		return true
+	}
+
+	for _, cve := range r.ScannedCves {
+		if 0 < len(cve.CveContents) {
+			return false
+		}
+	}
+	return true
+}
+
+func overwriteJSONFile(dir string, r models.ScanResult) error {
+	before := config.Conf.FormatJSON
+	beforeDiff := config.Conf.Diff
+	config.Conf.FormatJSON = true
+	config.Conf.Diff = false
+	w := LocalFileWriter{CurrentDir: dir}
+	if err := w.Write(r); err != nil {
+		return fmt.Errorf("Failed to write summary report: %s", err)
+	}
+	config.Conf.FormatJSON = before
+	config.Conf.Diff = beforeDiff
+	return nil
+}
+
+func loadPrevious(current models.ScanResults) (previous models.ScanResults, err error) {
+	dirs, err := ListValidJSONDirs()
+	if err != nil {
+		return
+	}
+
+	for _, result := range current {
+		for _, dir := range dirs[1:] {
+			var r models.ScanResult
+			path := filepath.Join(dir, result.ServerName+".json")
+			if r, err = loadOneServerScanResult(path); err != nil {
+				continue
+			}
+			if r.Family == result.Family && r.Release == result.Release {
+				previous = append(previous, r)
+				util.Log.Infof("Privious json found: %s", path)
+				break
+			}
+		}
+	}
+	return previous, nil
+}
+
+func diff(curResults, preResults models.ScanResults) (diffed models.ScanResults, err error) {
+	for _, current := range curResults {
+		found := false
+		var previous models.ScanResult
+		for _, r := range preResults {
+			if current.ServerName == r.ServerName {
+				found = true
+				previous = r
+				break
+			}
+		}
+
+		if found {
+			current.ScannedCves = getDiffCves(previous, current)
+			packages := models.Packages{}
+			for _, s := range current.ScannedCves {
+				for _, name := range s.PackageNames {
+					p := current.Packages[name]
+					packages[name] = p
+				}
+			}
+			current.Packages = packages
+		}
+
+		diffed = append(diffed, current)
+	}
+	return diffed, err
+}
+
+func getDiffCves(previous, current models.ScanResult) models.VulnInfos {
+	previousCveIDsSet := map[string]bool{}
+	for _, previousVulnInfo := range previous.ScannedCves {
+		previousCveIDsSet[previousVulnInfo.CveID] = true
+	}
+
+	new := models.VulnInfos{}
+	updated := models.VulnInfos{}
+	for _, v := range current.ScannedCves {
+		if previousCveIDsSet[v.CveID] {
+			if isCveInfoUpdated(v.CveID, previous, current) {
+				updated[v.CveID] = v
+			}
+		} else {
+			new[v.CveID] = v
+		}
+	}
+
+	for cveID, vuln := range new {
+		updated[cveID] = vuln
+	}
+	return updated
+}
+
+func isCveInfoUpdated(cveID string, previous, current models.ScanResult) bool {
+	cTypes := []models.CveContentType{
+		models.NVD,
+		models.JVN,
+		models.NewCveContentType(current.Family),
+	}
+
+	prevLastModified := map[models.CveContentType]time.Time{}
+	for _, c := range previous.ScannedCves {
+		if cveID == c.CveID {
+			for _, cType := range cTypes {
+				content, _ := c.CveContents[cType]
+				prevLastModified[cType] = content.LastModified
+			}
+			break
+		}
+	}
+
+	curLastModified := map[models.CveContentType]time.Time{}
+	for _, c := range current.ScannedCves {
+		if cveID == c.CveID {
+			for _, cType := range cTypes {
+				content, _ := c.CveContents[cType]
+				curLastModified[cType] = content.LastModified
+			}
+			break
+		}
+	}
+	for _, cType := range cTypes {
+		if equal := prevLastModified[cType].Equal(curLastModified[cType]); !equal {
+			return true
+		}
+	}
+	return false
+}
+
+// jsonDirPattern is file name pattern of JSON directory
+// 2016-11-16T10:43:28+09:00
+// 2016-11-16T10:43:28Z
+var jsonDirPattern = regexp.MustCompile(
+	`^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}(?:Z|[+-]\d{2}:\d{2})$`)
+
+// ListValidJSONDirs returns valid json directory as array
+// Returned array is sorted so that recent directories are at the head
+func ListValidJSONDirs() (dirs []string, err error) {
+	var dirInfo []os.FileInfo
+	if dirInfo, err = ioutil.ReadDir(config.Conf.ResultsDir); err != nil {
+		err = fmt.Errorf("Failed to read %s: %s",
+			config.Conf.ResultsDir, err)
+		return
+	}
+	for _, d := range dirInfo {
+		if d.IsDir() && jsonDirPattern.MatchString(d.Name()) {
+			jsonDir := filepath.Join(config.Conf.ResultsDir, d.Name())
+			dirs = append(dirs, jsonDir)
+		}
+	}
+	sort.Slice(dirs, func(i, j int) bool {
+		return dirs[j] < dirs[i]
+	})
+	return
+}
+
+// JSONDir returns
+// If there is an arg, check if it is a valid format and return the corresponding path under results.
+// If arg passed via PIPE (such as history subcommand), return that path.
+// Otherwise, returns the path of the latest directory
+func JSONDir(args []string) (string, error) {
+	var err error
+	dirs := []string{}
+
+	if 0 < len(args) {
+		if dirs, err = ListValidJSONDirs(); err != nil {
+			return "", err
+		}
+
+		path := filepath.Join(config.Conf.ResultsDir, args[0])
+		for _, d := range dirs {
+			ss := strings.Split(d, string(os.PathSeparator))
+			timedir := ss[len(ss)-1]
+			if timedir == args[0] {
+				return path, nil
+			}
+		}
+
+		return "", fmt.Errorf("Invalid path: %s", path)
+	}
+
+	// PIPE
+	if config.Conf.Pipe {
+		bytes, err := ioutil.ReadAll(os.Stdin)
+		if err != nil {
+			return "", fmt.Errorf("Failed to read stdin: %s", err)
+		}
+		fields := strings.Fields(string(bytes))
+		if 0 < len(fields) {
+			return filepath.Join(config.Conf.ResultsDir, fields[0]), nil
+		}
+		return "", fmt.Errorf("Stdin is invalid: %s", string(bytes))
+	}
+
+	// returns latest dir when no args or no PIPE
+	if dirs, err = ListValidJSONDirs(); err != nil {
+		return "", err
+	}
+	if len(dirs) == 0 {
+		return "", fmt.Errorf("No results under %s",
+			config.Conf.ResultsDir)
+	}
+	return dirs[0], nil
+}
+
+// LoadScanResults read JSON data
+func LoadScanResults(jsonDir string) (results models.ScanResults, err error) {
+	var files []os.FileInfo
+	if files, err = ioutil.ReadDir(jsonDir); err != nil {
+		return nil, fmt.Errorf("Failed to read %s: %s", jsonDir, err)
+	}
+	for _, f := range files {
+		if filepath.Ext(f.Name()) != ".json" || strings.HasSuffix(f.Name(), "_diff.json") {
+			continue
+		}
+
+		var r models.ScanResult
+		path := filepath.Join(jsonDir, f.Name())
+		if r, err = loadOneServerScanResult(path); err != nil {
+			return nil, err
+		}
+
+		results = append(results, r)
+	}
+	if len(results) == 0 {
+		return nil, fmt.Errorf("There is no json file under %s", jsonDir)
+	}
+	return
+}
+
+// loadOneServerScanResult read JSON data of one server
+func loadOneServerScanResult(jsonFile string) (result models.ScanResult, err error) {
+	var data []byte
+	if data, err = ioutil.ReadFile(jsonFile); err != nil {
+		err = fmt.Errorf("Failed to read %s: %s", jsonFile, err)
+		return
+	}
+	if json.Unmarshal(data, &result) != nil {
+		err = fmt.Errorf("Failed to parse %s: %s", jsonFile, err)
+	}
+	return
 }
